Checkpoint MLX array porting work

+++ lib/matplotlib/tests/test_ticker.py
@@ -980,7 +980,7 @@ class TestLogFormatterExponent:
         # formatted.
         (False, 10, mx.array([0.1, 0.00001, mx.pi, 0.2, -0.2, -0.00001]),
          range(6), ['0.1', '1e-05', '3.14', '0.2', '-0.2', '-1e-05']),
-        (False, 50, mx.array([3, 5, 12, 42], dtype=float), range(6),
+        (False, 50, mx.array([3, 5, 12, 42], dtype=mx.float32), range(6),
          ['3', '5', '12', '42']),
     ]
